+++ x/checkpointing/types/msgs.go
@@ -7,8 +7,10 @@ import (
 	ed255192 "github.com/cosmos/cosmos-sdk/crypto/keys/ed25519"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	authcodec "github.com/cosmos/cosmos-sdk/x/auth/codec"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
+	appparams "github.com/babylonlabs-io/babylon/v3/app/params"
 	"github.com/babylonlabs-io/babylon/v3/crypto/bls12381"
 )
 
@@ -36,6 +38,20 @@ func (m *MsgWrappedCreateValidator) ValidateBasic() error {
 	if m.MsgCreateValidator == nil {
 		return errors.New("MsgCreateValidator is nil")
 	}
+
+	if m.Key == nil {
+		return errors.New("BLS key is nil")
+	}
+
+	cdc := authcodec.NewBech32Codec(appparams.Bech32PrefixValAddr)
+	if err := m.MsgCreateValidator.Validate(cdc); err != nil {
+		return err
+	}
+
+	if err := m.Key.ValidateBasic(); err != nil {
+		return err
+	}
+
 	var pubKey ed255192.PubKey
 	err := pubKey.Unmarshal(m.MsgCreateValidator.Pubkey.GetValue())
 	if err != nil {