+++ x/finality/types/msg.go
@@ -1,10 +1,12 @@
 package types
 
 import (
-	errorsmod "cosmossdk.io/errors"
 	"fmt"
+
+	errorsmod "cosmossdk.io/errors"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
+	bbntypes "github.com/babylonlabs-io/babylon/v3/types"
 	"github.com/cometbft/cometbft/crypto/merkle"
 	"github.com/cometbft/cometbft/crypto/tmhash"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -19,6 +21,12 @@ var (
 	_ sdk.Msg = &MsgUpdateParams{}
 	_ sdk.Msg = &MsgAddFinalitySig{}
 	_ sdk.Msg = &MsgCommitPubRandList{}
+	_ sdk.Msg = &MsgUnjailFinalityProvider{}
+	// Ensure msgs implement ValidateBasic
+	_ sdk.HasValidateBasic = &MsgAddFinalitySig{}
+	_ sdk.HasValidateBasic = &MsgResumeFinalityProposal{}
+	_ sdk.HasValidateBasic = &MsgCommitPubRandList{}
+	_ sdk.HasValidateBasic = &MsgUnjailFinalityProvider{}
 )
 
 const ExpectedCommitmentLengthBytes = 32
@@ -28,8 +36,20 @@ func (m *MsgAddFinalitySig) MsgToSign() []byte {
 }
 
 func (m *MsgAddFinalitySig) ValidateBasic() error {
+	if _, err := sdk.AccAddressFromBech32(m.Signer); err != nil {
+		return sdkerrors.ErrInvalidAddress.Wrapf("invalid signer address: got %s", m.Signer)
+	}
+
+	if m.FpBtcPk == nil {
+		return ErrInvalidFinalitySig.Wrap("empty Finality Provider BTC PubKey")
+	}
+
 	if m.FpBtcPk.Size() != bbn.BIP340PubKeyLen {
 		return ErrInvalidFinalitySig.Wrapf("invalid finality provider BTC public key length: got %d, want %d", m.FpBtcPk.Size(), bbn.BIP340PubKeyLen)
+	}
+
+	if m.PubRand == nil {
+		return ErrInvalidFinalitySig.Wrap("empty Public Randomness")
 	}
 
 	if m.PubRand.Size() != bbn.SchnorrPubRandLen {
@@ -38,6 +58,10 @@ func (m *MsgAddFinalitySig) ValidateBasic() error {
 
 	if m.Proof == nil {
 		return ErrInvalidFinalitySig.Wrap("empty inclusion proof")
+	}
+
+	if m.FinalitySig == nil {
+		return ErrInvalidFinalitySig.Wrap("empty finality signature")
 	}
 
 	if m.FinalitySig.Size() != bbn.SchnorrEOTSSigLen {
@@ -104,6 +128,9 @@ func (m *MsgCommitPubRandList) VerifySig() error {
 	if err != nil {
 		return err
 	}
+	if m.FpBtcPk == nil {
+		return fmt.Errorf("empty FP BTC PubKey")
+	}
 	pk, err := m.FpBtcPk.ToBTCPK()
 	if err != nil {
 		return err
@@ -128,6 +155,10 @@ func (m *MsgCommitPubRandList) ValidateBasic() error {
 		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid signer address (%s)", err)
 	}
 
+	if m.FpBtcPk == nil {
+		return ErrInvalidPubRand.Wrap("empty FP BTC PubKey")
+	}
+
 	// Checks if the commitment is exactly 32 bytes
 	if len(m.Commitment) != ExpectedCommitmentLengthBytes {
 		return ErrInvalidPubRand.Wrapf("commitment must be %d bytes, got %d", ExpectedCommitmentLengthBytes, len(m.Commitment))
@@ -140,6 +171,50 @@ func (m *MsgCommitPubRandList) ValidateBasic() error {
 			"public rand commit start block height: %d is equal or higher than (start height + num pub rand) %d",
 			m.StartHeight, m.StartHeight+m.NumPubRand,
 		)
+	}
+
+	if m.Sig == nil {
+		return ErrInvalidPubRand.Wrap("empty signature")
+	}
+
+	return nil
+}
+
+// ValidateBasic performs stateless validation on MsgUnjailFinalityProvider
+func (m *MsgUnjailFinalityProvider) ValidateBasic() error {
+	if _, err := sdk.AccAddressFromBech32(m.Signer); err != nil {
+		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid signer address (%s)", err)
+	}
+	if m.FpBtcPk == nil {
+		return ErrInvalidUnjailFinalityProvider.Wrap("empty FP BTC PubKey")
+	}
+
+	return nil
+}
+
+func (m *MsgResumeFinalityProposal) ValidateBasic() error {
+	if _, err := sdk.AccAddressFromBech32(m.Authority); err != nil {
+		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid authority address (%s)", err)
+	}
+	if m.HaltingHeight == 0 {
+		return ErrInvalidResumeFinality.Wrap("halting height is zero")
+	}
+	if len(m.FpPksHex) == 0 {
+		return ErrInvalidResumeFinality.Wrap("no fp pk hex set")
+	}
+
+	fps := make(map[string]struct{})
+	for _, fpPkHex := range m.FpPksHex {
+		_, err := bbntypes.NewBIP340PubKeyFromHex(fpPkHex)
+		if err != nil {
+			return ErrInvalidResumeFinality.Wrapf("failed to parse FP BTC PK Hex (%s) into BIP-340", fpPkHex)
+		}
+
+		_, found := fps[fpPkHex]
+		if found {
+			return ErrInvalidResumeFinality.Wrapf("duplicated FP BTC PK Hex (%s)", fpPkHex)
+		}
+		fps[fpPkHex] = struct{}{}
 	}
 
 	return nil